merge'cota/release'into n7.1
The re-upbringing of COTA (CypherOS OTA)
*Cough* But we don't code though ;)

Special thanks to Aaron Nixon (CypheOS Team Member)
for his integration of global tracking. Also thanks
to our translators for making COTA available for world-wide.

The application is based on an older version of ParanoidHub,
completely rewritten. Special thanks to PA for their update
services, and providing the the base that made COTA excellent.