Cota: Set ProgressBar in it's correct slot
Previously the ProgressBar's visibilty was removed from the main activity,
causing it to appear in any usage case. Let's add this back so the system
knows when the Progressbar is supposed to initiate. Also move it under the
header message instead to stay out of the Button and Update Sizes way.

Also let's ensure the system know, the button is below the message, not causing
it to remain static in some cases and let's set the notification to be created
dynamically.

Change-Id: I6e7be030e10c51485d08b6ad6dab736759d9d9fc