Cota: Cleanup logic at runtime
Use Google's approach and don't show the DownloadManager
notification while the download is in progress. This progress
will be set within the app. Once the download complete's
the user will be generated with a new notification verifying
a completed download and prompting for install.

Fixed issue where the update notification shows after the package
is still being downloaded. Also removed progressbar. We'll stick
with percentage and keep it simple.

Change-Id: I3faf868fe50dc5c7ef9fc40f98f173b455a06bbf