Cota: Inherit scrollview for update description
Instead of including our dynamic update messages with
static text, let's branch this out into it's own
scroll based textview. This way we can create a markdown
of changes and/or push more information to the user without
crowding the main layout.

Also make sure some views are gone instead of invisible and
allow layout changes to be animated

Change-Id: I38a27d28cff687394d59e54c2ef1a84729a5c028